requirements.txt
pyyaml>=6.0.1
pydantic>=2.5.2
psycopg2-binary>=2.9.9  # Para PostgreSQL 